ir: Completely redo and add JSON Typedef schema
JSON Typedef[1] was chosen rather than the older JSON Schema[2]
because it's a significantly simpler language that is also more
directly focused on being codegen-able.
JSON Typedef appears to be rather unmaintained, but a project
called `ajv`[3] created the image that it was more standardized and
maintained than it actually was.

[1]: https://jsontypedef.com/
[2]: https://json-schema.org/
[3]: https://ajv.js.org/json-type-definition.html