Make bool parse/encode stable with panic
Fixes https://github.com/gtker/wow_messages/issues/84.

On a trial basis values on than 0/1 will panic for bools.
This should provide opportunity for finding fields that either aren't
really bools, or where some other value is used for true/false.

If fields are found that use e.g. 0/7 as false/true, then it might make
sense to either change it to an enum or add a tag to specify it.